Modify the data type in the ping command file
MIRTOS-468

+++ system/ping/ping.c
@@ -51,8 +51,8 @@
 struct ping_priv_s
 {
   int code;                      /* Notice code ICMP_I/E/W_XXX */
-  int tmin;                      /* Minimum round trip time */
-  int tmax;                      /* Maximum round trip time */
+  long tmin;                     /* Minimum round trip time */
+  long tmax;                     /* Maximum round trip time */
   long long tsum;                /* Sum of all times, for doing average */
   long long tsum2;               /* Sum2 is the sum of the squares of sum ,for doing mean deviation */
 };
@@ -125,7 +125,7 @@ static void ping_result(FAR const struct ping_result_s *result)
         break;
 
       case ICMP_E_SOCKET:
-        fprintf(stderr, "ERROR: socket() failed: %d\n", result->extra);
+        fprintf(stderr, "ERROR: socket() failed: %ld\n", result->extra);
         break;
 
       case ICMP_I_BEGIN:
@@ -138,21 +138,21 @@ static void ping_result(FAR const struct ping_result_s *result)
         break;
 
       case ICMP_E_SENDTO:
-        fprintf(stderr, "ERROR: sendto failed at seqno %u: %d\n",
+        fprintf(stderr, "ERROR: sendto failed at seqno %u: %ld\n",
                 result->seqno, result->extra);
         break;
 
       case ICMP_E_SENDSMALL:
-        fprintf(stderr, "ERROR: sendto returned %d, expected %u\n",
+        fprintf(stderr, "ERROR: sendto returned %ld, expected %u\n",
                 result->extra, result->outsize);
         break;
 
       case ICMP_E_POLL:
-        fprintf(stderr, "ERROR: poll failed: %d\n", result->extra);
+        fprintf(stderr, "ERROR: poll failed: %ld\n", result->extra);
         break;
 
       case ICMP_W_TIMEOUT:
-        printf("No response from %u.%u.%u.%u: icmp_seq=%u time=%d ms\n",
+        printf("No response from %u.%u.%u.%u: icmp_seq=%u time=%ld ms\n",
                (unsigned int)(result->dest.s_addr) & 0xff,
                (unsigned int)(result->dest.s_addr >> 8) & 0xff,
                (unsigned int)(result->dest.s_addr >> 16) & 0xff,
@@ -161,23 +161,23 @@ static void ping_result(FAR const struct ping_result_s *result)
         break;
 
       case ICMP_E_RECVFROM:
-        fprintf(stderr, "ERROR: recvfrom failed: %d\n", result->extra);
+        fprintf(stderr, "ERROR: recvfrom failed: %ld\n", result->extra);
         break;
 
       case ICMP_E_RECVSMALL:
-        fprintf(stderr, "ERROR: short ICMP packet: %d\n", result->extra);
+        fprintf(stderr, "ERROR: short ICMP packet: %ld\n", result->extra);
         break;
 
       case ICMP_W_IDDIFF:
         fprintf(stderr,
-                "WARNING: Ignoring ICMP reply with ID %d.  "
+                "WARNING: Ignoring ICMP reply with ID %ld.  "
                 "Expected %u\n",
                 result->extra, result->id);
         break;
 
       case ICMP_W_SEQNOBIG:
         fprintf(stderr,
-                "WARNING: Ignoring ICMP reply to sequence %d.  "
+                "WARNING: Ignoring ICMP reply to sequence %ld.  "
                 "Expected <= %u\n",
                 result->extra, result->seqno);
         break;
@@ -199,7 +199,7 @@ static void ping_result(FAR const struct ping_result_s *result)
             priv->tmax = result->extra;
           }
 
-        printf("%u bytes from %u.%u.%u.%u: icmp_seq=%u time=%d.%d ms\n",
+        printf("%u bytes from %u.%u.%u.%u: icmp_seq=%u time=%ld.%ld ms\n",
                result->info->datalen,
                (unsigned int)(result->dest.s_addr) & 0xff,
                (unsigned int)(result->dest.s_addr >> 8) & 0xff,
@@ -212,7 +212,7 @@ static void ping_result(FAR const struct ping_result_s *result)
       case ICMP_W_RECVBIG:
         fprintf(stderr,
                 "WARNING: Ignoring ICMP reply with different payload "
-                "size: %d vs %u\n",
+                "size: %ld vs %u\n",
                 result->extra, result->outsize);
         break;
 
@@ -221,7 +221,7 @@ static void ping_result(FAR const struct ping_result_s *result)
         break;
 
       case ICMP_W_TYPE:
-        fprintf(stderr, "WARNING: ICMP packet with unknown type: %d\n",
+        fprintf(stderr, "WARNING: ICMP packet with unknown type: %ld\n",
                 result->extra);
         break;
 
@@ -237,17 +237,16 @@ static void ping_result(FAR const struct ping_result_s *result)
                    result->nrequests;
 
             printf("%u packets transmitted, %u received, %u%% packet loss, "
-                   "time %d ms\n",
+                   "time %ld ms\n",
                    result->nrequests, result->nreplies, tmp,
                    result->extra / USEC_PER_MSEC);
             if (result->nreplies > 0)
               {
-                int avg = priv->tsum / result->nreplies;
-                int tmdev = sqrt(priv->tsum2 / result->nreplies -
+                long avg = priv->tsum / result->nreplies;
+                long tmdev = sqrt(priv->tsum2 / result->nreplies -
                                  (long long)avg * avg);
-
-                printf("rtt min/avg/max/mdev = %d.%03d/%u.%03u/"
-                       "%d.%03d/%d.%03d ms\n",
+                printf("rtt min/avg/max/mdev = %ld.%03ld/%ld.%03ld/"
+                       "%ld.%03ld/%ld.%03ld ms\n",
                        priv->tmin / USEC_PER_MSEC,
                        priv->tmin % USEC_PER_MSEC,
                        avg / USEC_PER_MSEC, avg % USEC_PER_MSEC,
